# Cache function in scalar model generate