Before this PR:
```
julia> solve_scs_moi(data, vector_version = false, max_iters = 1)
-7.320596134944753e6

julia> TimerOutputs.get_defaulttimer()
 ─────────────────────────────────────────────────────────────────────────
                                  Time                   Allocations      
                          ──────────────────────   ───────────────────────
     Tot / % measured:         5.41s / 15.0%            322MiB / 100%     

 Section          ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────
 SCS MOI scalar        1    808ms   100%   808ms    322MiB  100%    322MiB
   generate            1    566ms  70.1%   566ms    238MiB  73.9%   238MiB
   solve               1    242ms  29.9%   242ms   83.8MiB  26.0%  83.8MiB
 ─────────────────────────────────────────────────────────────────────────
```
After this PR:
```
julia> solve_scs_moi(data, vector_version = false, max_iters = 1)
-8.575051909370726e6

julia> TimerOutputs.get_defaulttimer()
 ─────────────────────────────────────────────────────────────────────────
                                  Time                   Allocations      
                          ──────────────────────   ───────────────────────
     Tot / % measured:         3.24s / 26.8%            305MiB / 100%     

 Section          ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────
 SCS MOI scalar        1    870ms   100%   870ms    305MiB  100%    305MiB
   generate            1    580ms  66.6%   580ms    221MiB  72.5%   221MiB
   solve               1    290ms  33.4%   290ms   83.8MiB  27.5%  83.8MiB
 ─────────────────────────────────────────────────────────────────────────
```
The timings are not so stable but the allocations are. I'm surprised we don't gain more on allocations for `generate` by the way.