@odow 2020-09-02
What `data` did you use? When I ran this, I got essentially no change in the full latex table. I'm in favor of closing.